LPS-179756 Extract LayoutPageTemplateEntryCPDefinitionDisplayLayoutSelectionScreenNavigationCategory methods to a LayoutPageTemplateEntryCPDefinitionDisplayLayoutSelectionScreenNavigationEntry component and extends LayoutPageTemplateEntryCPDefinitionDisplayLayoutSelectionScreenNavigationCategory in it
